requirements.txt
pip install opencv-python pillow pyautogui
pip install pytesseract
opencv-python~=4.12.0.88
numpy~=2.2.6
PyAutoGUI~=0.9.54
pillow~=10.3.0
pytesseract~=0.3.13